Add option to create installer
I added the option --binary-only to satisfy the base requirement of this
ticket, which is to allow users to avoid creating the installer if they
want. I don't like passing through negative options in a system - it's a
recipe for boolean errors. So instead we leave the default in place - to
build the installer - and use an option to avoid it - binary-only. We then
negate the logic as soon as we are away from the high-level user interface
and instead pass through the much more clear 'create_installer' option

I did some formatting in all of this too because I think it makes all this
code more readable and extensible