Add low-level support for not building an installer
This ended up being a larger commit than I would have liked, but in adding
the property to track that we wanted to build the package I realized we
could just remove the verbose property as well. That meant I had to remove
the plumbing for it through the rest of the system, including
installer_mac. That was easy enough, but it makes the commit large.

In both installer_mac and installer_win I opted to invert the error
conditions so that we can save on indentation and exit early