Get more explicit about requirements to build launchwin
I worked a bunch on understanding how the launcher works and gets built.
In doing so I discovered that I can't build it with python 3.5 because
we're using py2exe which doesn't support python 3.5. I also discovered
there's some requirements that aren't explicitly installed and there's not
a good way to install them on behalf of the user. So, at the least we can
let them know about the problems and make it easier for them to get the
right output and decent errors

This also changes the launch builder to take in a parameter that is the
location of the python3.4 executable. It defaults to the default location
of the install on windows