Add support for both window and console installations
This only really changes windows. Now when we build the launchers we
create a launcher that supports the console as well as one that just
supports windows applications. This is delimited now by the variant and
means we have 4 different launcher packages for windows, x86 and amd64 in
the console and window variants

I couldn't find a good way of sharing the launcher code between the two so
I created a separate launcher for the console. The only difference is that
it uses 'python.exe' rather than 'pythonw.exe'