Update launchers to 0.0.5
In this update we now have a more sophisticated system for looking
for the osnap directory. Previously we looked based on the current
working directory, which is mostly fine, except it doesn't allow for
running programs from other locations. We might do this if we have a
service or a console application